RequirementAttribute has now 2 constructors:
The original:
public RequirementAttribute(string documentation)

And a new one:
public RequirementAttribute(string id, string documentation)

The original constructor is adding now a default id to a requirement (static counter) 
so even previous projects can use the new feature and 
get a list of requirements that executed by each test in the ct log.

(In order to get requirements coverage you must supply the requirements external file and 
have the ids in-synch with the ids of the RequirementAttribute instances in the model).